// src/models/command/lib/BranchType.ts
/**
 * The type of the branch.
 */
export enum BranchType {
  // Starts Normal path, background is the default grey. Represented by `#N`.
  Normal = 'N',

  // Starts Advanced or Professional path, background is blue. Represented by `#E`.
  Advanced = 'E',

  // Starts Expert path, background is purple. Represented by `#M`.
  Master = 'M',
}